[common/datablocks] move functions to kernels
