+++ common/datablocks/src/kernels/mod.rs
@@ -1,0 +1,19 @@
+// Copyright 2020-2021 The Datafuse Authors.
+//
+// SPDX-License-Identifier: Apache-2.0.
+
+#[cfg(test)]
+mod data_block_concat_test;
+#[cfg(test)]
+mod data_block_scatter_test;
+#[cfg(test)]
+mod data_block_sort_test;
+#[cfg(test)]
+mod data_block_take_test;
+
+mod data_block_concat;
+mod data_block_scatter;
+mod data_block_sort;
+mod data_block_take;
+
+pub use data_block_sort::SortColumnDescription;
