+++ common/datablocks/src/lib.rs
@@ -3,14 +3,12 @@
 // SPDX-License-Identifier: Apache-2.0.
 
 #[cfg(test)]
-mod data_block_kernel_test;
-#[cfg(test)]
 mod data_block_test;
 
 mod data_block;
 mod data_block_debug;
-mod data_block_kernel;
+mod kernels;
 
 pub use data_block::DataBlock;
 pub use data_block_debug::*;
-pub use data_block_kernel::*;
+pub use kernels::SortColumnDescription;
